fix apple silicon builds
The previous fix for apple silicon builds was not correctly configuring CMake to build for the correct target architecture. The approach implemented here patterns off of that used in [parselmouth](https://github.com/YannickJadoul/Parselmouth/blob/master/.github/workflows/wheels.yml) in which the matrix of build targets is manually configured so that the correct environment variables can be specified at run time. In particular `CMAKE_OSX_ARCHITECTURES` must be set so that CMake builds for the correct architecture.

This approach is a little awkward, I would expect CIBuildWheel to be updated so that it can do this for us. As noted by the robust-laplacian package maintainer, other build systems such as scikit-build already do this correctly.